Refactor /namegen to MVC
